# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Zooot VPN API
  version: 0.2.0
paths:
  /api/v1/auth/register:
    post:
      summary: Register user
      requestBody:
        required: true
      responses:
        '201': { description: Created }
        '400': { description: VALIDATION_ERROR }
        '409': { description: USER_ALREADY_EXISTS }
  /api/v1/auth/login:
    post:
      summary: Login user
      responses:
        '200': { description: OK }
        '400': { description: VALIDATION_ERROR }
        '401': { description: INVALID_CREDENTIALS }
  /api/v1/config/resolve-token:
    post:
      summary: Resolve config token with subscription and servers
      responses:
        '200':
          description: Resolved
          content:
            application/json:
              schema:
                type: object
                properties:
                  servers:
                    type: array
                    items:
                      type: object
                      properties:
                        protocols:
                          type: array
                          items:
                            type: object
                            properties:
                              type:
                                type: string
                                enum: [wireguard, xray_vless_reality, amneziawg, openvpn_udp, openvpn_tcp]
                              port:
                                type: integer
                              config:
                                type: string
                                nullable: true
                                description: Non-empty for provisioned WireGuard or Xray VLESS Reality client configs.
                              config_source:
                                type: string
                                nullable: true
        '400': { description: VALIDATION_ERROR }
        '403': { description: SUBSCRIPTION_INACTIVE }
        '404': { description: CONFIG_TOKEN_NOT_FOUND }
  /api/v1/servers/recommended:
    get:
      summary: Get online recommended servers and healthy protocols
      responses:
        '200': { description: OK }
components:
  schemas:
    ApiError:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string